move recurrence validation to RecurrenceValidator
