+++ src/Microsoft.FeatureManagement/FeatureFilters/Recurrence/RecurrenceRange.cs
@@ -23,6 +23,6 @@ namespace Microsoft.FeatureManagement.FeatureFilters
         /// <summary>
         /// The number of times to repeat the time window.
         /// </summary>
-        public int NumberOfOccurrences { get; set; } = 1;
+        public int NumberOfOccurrences { get; set; } = int.MaxValue;
     }
 }
